Separate commands section in one separated module
